requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
multidict==6.0.4
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
sqlparse==0.4.4
twilio==8.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
yarl==1.9.2
